backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
instructor
groq
pydantic
eval_type_backport
pyjwt
cryptography
httpx
